# gofr/swagger.yaml
swagger: '2.0'
info:
  description: Applying crud operations on Books and Authors dataset.
  version: '1.0'
  title: library
  contact:
    email: suhani.singh@zopsmart.com
host: localhost:8000
tags:
  - name: Book
    description: Details about the book
  - name: Author
    description: Details about the Author
schemes:
  - http
paths:
  /book:
    get:
      tags:
        - Book
      summary: Get books details
      description: Fetches the book details
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: title
          in: query
          description: Return Book details with title
          required: false
          type: string
          format: string
        - name: includeAuthor
          in: query
          description: Return Author values
          required: false
          type: boolean
          format: string
      responses:
        '200':
          description: data found successfully
          schema:
            $ref: '#/definitions/Book'
        '500':
          description: Internal Server Error
    post:
      tags:
        - Book
      summary: Create a new Book
      description: It adds a new Book to the database
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: Creates a Book object
          required: true
          schema:
            $ref: '#/definitions/Book'
      responses:
        '201':
          description: Book created successfully
          schema:
            $ref: '#/definitions/Book'
        '400':
          description: Bad Request
        '409':
          description: Status Conflict
        '500':
          description: Internal Server Error
  /author:
    post:
      tags:
        - Author
      summary: Create a new Author
      description: It adds a new Author to the database
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: Creates a Author object
          required: true
          schema:
            $ref: '#/definitions/Author'
      responses:
        '201':
          description: Author created successfully
          schema:
            $ref: '#/definitions/Author'
        '400':
          description: Bad Request
        '409':
          description: Status Conflict
        '500':
          description: Internal Server Error
  /book/{id}:
    get:
      tags:
        - Book
      summary: GET and Prints details of the Book by id
      description: Prints the details of the Book by id
      operationId: Get
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: ID of book to get the details
          required: true
          type: string
          format: string
      responses:
        '200':
          description: Data fetched
          schema:
            $ref: '#/definitions/Book'
        '400':
          description: Bad Request
        '404':
          description: No entry found
        '500':
          description: Internal Server Error
    put:
      tags:
        - Book
      summary: Update book by id
      description: Update the book details entered by user
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: ID of book to update
          required: true
          type: string
          format: string
      responses:
        '200':
          description: Successfully updated
          schema:
            $ref: '#/definitions/Book'
        '400':
          description: Bad Request
        '404':
          description: No entry updated
        '500':
          description: Internal Server Error
    delete:
      tags:
        - Book
      summary: Deletes the book by id
      description: ''
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: ID of book to delete
          required: true
          type: string
          format: string
      responses:
        '204':
          description: No content successful
        '400':
          description: Bad Request
        '404':
          description: No entry deleted
        '500':
          description: Internal Server Error
  /author/{id}:
    put:
      tags:
        - Author
      summary: Update Author by id
      description: Update the Author details entered by user
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: ID of Author to update
          required: true
          type: string
          format: string
      responses:
        '200':
          description: Successfully updated
          schema:
            $ref: '#/definitions/Book'
        '400':
          description: Bad Request
        '404':
          description: No entry updated
        '500':
          description: Internal Server Error
    delete:
      tags:
        - Author
      summary: Deletes the Author by id
      description: ''
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: ID of Author to delete
          required: true
          type: string
          format: string
      responses:
        '204':
          description: No content successful
        '404':
          description: No entry deleted
        '500':
          description: Internal Server Error
definitions:
  Book:
    type: object
    properties:
      ID:
        type: integer
        format: int64
      Title:
        type: string
        format: string
      Publication:
        type: string
        description: Publication
        enum:
          - Arihanth
          - Scholastic
          - Penguin
      PublishedDate:
        type: string
        description: Date of Pulication
        format: DD/MM/YYYY
      Author:
        $ref: '#/definitions/Author'
  Author:
    type: object
    properties:
      ID:
        type: integer
        format: int64
      FirstName:
        type: string
        format: string
      LastName:
        type: string
        format: string
      Dob:
        type: string
        format: DD/MM/YYYY
      PenName:
        type: string
        format: string
externalDocs:
  description: ''
  url: https://github.com/Suhani-zs